\name{jagstools-package}
\alias{jagstools-package}
\alias{jagstools}
\docType{package}
\title{
Summarise/filter rjags/mcmc.list/stanfit results.
}
\description{
Functions for handling rjags, mcmc.list and stanfit objects:
\code{rhats} prints mean, sd and rhat for each parameter of an rjags object, with optional sorting
\code{jagsresults} prints parameter estimates etc. for specified parameters using regex pattern matching
\code{rearray} recovers array structure for multidimensional parameters
}
\details{
\tabular{ll}{
Package: \tab jagstools\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2015-03-11\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
John Baumgartner & James Camac
Maintainer: John Baumgartner <johnbaums@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
