% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simgrowth.R
\docType{data}
\name{simgrowth}
\alias{simgrowth}
\title{Simulated population growth model output}
\format{An \code{rjags} object with 102 estimated parameters.}
\usage{
simgrowth
}
\description{
An example \code{rjags} object including output for a hypothetical population
growth model fit to simulated data.
}
\author{
John Baumgartner, \email{johnbaums@gmail.com}
}
\keyword{datasets}

