\name{jagstools-package}
\alias{jagstools-package}
\alias{jagstools}
\docType{package}
\title{
Summarise and plot rjags/mcmc.list results.
}
\description{
Functions for handling rjags and mcmc.list objects:
\code{rhats} prints mean, sd and rhat for each parameter of an rjags object, with optional sorting
\code{jagsresults} prints parameter estimates etc. for specified parameters using regex pattern matching
\code{rearray} recovers array structure for multidimensional parameters
}
\details{
\tabular{ll}{
Package: \tab jagstools\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2012-11-25\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
John Baumgartner
Maintainer: John Baumgartner <johnbb@student.unimelb.edu.au>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
