\name{jagstools-package}
\alias{jagstools-package}
\alias{jagstools}
\docType{package}
\title{
Summarise and plot rjags results.
}
\description{
Functions for handling rjags object:
\code{rhats} prints mean, sd and rhat for each parameter, sorted by rhat (by default, but sorting can be disabled)
\code{jagsresults} prints parameter estimates etc. for specified parameters using regex pattern matching
}
\details{
\tabular{ll}{
Package: \tab jagstools\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2011-08-10\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
}
\author{
John Baumgartner
Maintainer: John Baumgartner <johnbb@student.unimelb.edu.au>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
