% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearray.R
\name{rearray}
\alias{rearray}
\title{Recover array structure for multidimensional nodes}
\usage{
rearray(x, param = "all", fields = "all", regex = TRUE, exact = FALSE)
}
\arguments{
\item{x}{The \code{rjags}, \code{rjags.parallel}, or \code{mcmc.list} object
for which array structure will be recovered.}

\item{param}{Character vector. The multidimensional parameters that will be 
given an array structure. Default is \code{'all'}, which returns arrays for
all multidimensional parameters in \code{x}.}

\item{fields}{A vector (character or integer) that indicates the summary 
fields that should be included as array slices. Valid fields are: 'mean', 
'sd', '2.5\%', '25\%', '50\%', '75\%', '97.5\%', 'Rhat', and 'n.eff'. If 
supplied as an integer vector, numbers should indicate the column numbers 
(i.e. positions in the series given above). Note that 'Rhat' and 'n.eff' 
are only available if \code{x} is an \code{rjags} object. Default is 
\code{'all'}, which returns all available fields as array slices.}

\item{regex}{Logical. If \code{regex} is \code{TRUE}, then \code{param} is 
expected to be a single string giving a text pattern to be matched. 
Parameters with names matching the pattern will be returned (unless 
\code{invert} is \code{TRUE}, which results in all parameters that do not 
match the pattern being returned). Text pattern matching uses regular 
expressions (\code{\link{regex}}).}

\item{exact}{Logical. If \code{exact} is \code{TRUE}, only parameters whose 
names match elements of \code{param} exactly will be returned. If 
\code{exact = FALSE}, parameters that have names containing the strings 
given by \code{param} will be returned. Ignored if \code{regex} is 
\code{TRUE}.}
}
\value{
A list of arrays (one per multidimensional model parameter specified
  in \code{param}) containing node summaries for the requested \code{fields}.
  Fields (e.g. mean, sd, Rhat, etc.) will be included as the last
  (right-most) dimension.
}
\description{
Converts summaries of multidimensional nodes of \code{rjags} or 
\code{mcmc.list} objects to a list of arrays.
}
\examples{
data(simgrowth)
a <- rearray(simgrowth)
str(a)
a2 <- rearray(simgrowth, param='lambda', fields=c('mean', '50\\\%', 'n.eff'))
str(a2)
# or...
a2 <- rearray(simgrowth, param='lambda', fields=c(1, 2, 9))
str(a2)
}
\author{
John Baumgartner, \email{johnbaums@gmail.com}
}
\seealso{
\code{\link{jagsresults}}
}

